Vital integration

+++ database/migrations/2023_04_18_120202_create_prices_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->foreignId('product')->constrained('articles');
             $table->date('debut');
             $table->date('fin');
-            $table->float('prix');
+            $table->decimal('prix', 10, 6);
             $table->boolean('active')->default(true);
             $table->timestamps();
         });
